_old_version/__requirements__.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
Flask-CORS==4.0.0
requests==2.31.0
openai==0.28.1
python-dotenv==1.0.0
Werkzeug==2.3.7
sentence-transformers==2.2.2
scikit-learn==1.3.0
numpy==1.24.3
torch==2.0.1